Revert to old behavior: LibgpiodDriver is V1, V2 is separate
