+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/RaspberryPi3LinuxDriver.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Buffers.Binary;
+using System.Device.Gpio.Drivers.Libgpiod;
 using System.Device.Gpio.Drivers.Libgpiod.V1;
 using System.Diagnostics;
 using System.IO;
@@ -628,7 +629,7 @@ internal unsafe class RaspberryPi3LinuxDriver : GpioDriver
     {
         try
         {
-            _interruptDriver = new LibGpiodV1Driver(0);
+            _interruptDriver = new LibGpiodDriver(0);
         }
         catch (PlatformNotSupportedException)
         {
