+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/UnixDriver.cs
@@ -1,6 +1,11 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Device.Gpio.Drivers.Libgpiod;
+using System.Device.Gpio.Drivers.Libgpiod.V2;
+using System.Device.Gpio.System.Device.Gpio.Drivers.Libgpiod.V1;
+using System.Diagnostics.CodeAnalysis;
+
 namespace System.Device.Gpio.Drivers;
 
 /// <summary>
@@ -32,18 +37,39 @@ public abstract class UnixDriver : GpioDriver
         }
 
         UnixDriver? driver = null;
-        try
+
+        if (TryCreate(() => new LibGpiodDriver(0), out driver))
         {
-#pragma warning disable SDGPIO0001 // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
-            driver =
-                (UnixDriver)LibGpiodDriverFactory.Instance.Create(0).LibGpiodDriver;
-#pragma warning restore SDGPIO0001 // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
-        }
-        catch (PlatformNotSupportedException)
-        {
-            driver = new SysFsDriver();
+            return driver;
         }
 
-        return driver;
+#pragma warning disable SDGPIO0001
+        if (TryCreate(() => new LibGpiodV2Driver(0), out driver))
+#pragma warning restore SDGPIO0001
+        {
+            return driver;
+        }
+
+        if (TryCreate(() => new SysFsDriver(), out driver))
+        {
+            return driver;
+        }
+
+        throw new PlatformNotSupportedException("No unix driver appears to be runnable");
+    }
+
+    private static bool TryCreate(Func<UnixDriver> creationAction, [NotNullWhen(true)]out UnixDriver? driver)
+    {
+        try
+        {
+            driver = creationAction();
+        }
+        catch (Exception x) when (x is PlatformNotSupportedException || x is DllNotFoundException)
+        {
+            driver = null;
+            return false;
+        }
+
+        return true;
     }
 }
